requirements.txt
# OpenCPO Gateway — Python Dependencies
# Python 3.11+ required (ships with Raspberry Pi OS Bookworm)

# ── Core ──────────────────────────────────────────────────────────────────────
websockets>=12.0           # OCPP WebSocket proxy
httpx>=0.27.0              # Async HTTP client (Core API, updates)
uvicorn[standard]>=0.29.0  # ASGI server for tap + troubleshoot APIs
fastapi>=0.111.0           # REST API framework
pyyaml>=6.0                # Config file parsing
cryptography>=42.0         # mTLS certs, key vault encryption
prometheus_client>=0.20.0  # Metrics exposition

# ── Sensor Array ─────────────────────────────────────────────────────────────
smbus2>=0.4.3                              # I2C bus access
RPi.GPIO>=0.7.1; platform_machine=="armv7l" or platform_machine=="aarch64"
gpiozero>=2.0                              # GPIO abstraction (Pi + mock)
adafruit-circuitpython-bme280>=2.6.0       # BME280 temp/humidity/pressure
adafruit-circuitpython-ads1x15>=2.3.0      # ADS1115 ADC (CT clamp)
adafruit-circuitpython-tsl2591>=1.3.0      # TSL2591 light sensor

# ── CCTV ──────────────────────────────────────────────────────────────────────
uiprotect>=5.0.0           # UniFi Protect API (same lib as Home Assistant)
Pillow>=10.0.0             # Lightweight image processing (motion detection)
numpy>=1.26.0              # Frame diff for motion detection

# ── Camera discovery ──────────────────────────────────────────────────────────
zeroconf>=0.131.0          # mDNS (charger discovery + UniFi Protect discovery)
# onvif-zeep>=0.2.12       # ONVIF camera support (optional, uncomment if needed)

# ── Optional / enhanced ───────────────────────────────────────────────────────
# aiofiles>=23.0           # Async file I/O for tap export
# wsdiscovery>=2.0         # WS-Discovery for ONVIF (pairs with onvif-zeep)
